+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-12-10
+
+### Changed
+- Correct handling of diposed instances of FileAndKeyEncryption and SplitKeyEncryption.
+- Throw ObjectDisposedException instead of InvalidOperationException if an instance of ProtectedByteArray or ShuffledByteArray is disposed of.
+
+## [1.3.1] - 2020-12-10
+
+### Changed
+- Made hashing simpler and 2.5 times faster.
+- Added missing test for source bytes being null.
+
 ## [1.3.0] - 2020-11-12
 
 ### Changed